Create isAdmin utility function
